+++ tests/lets.ml
@@ -1,6 +1,6 @@
 let n1 =
   let m = 12 in
-  let n1 = m in
+  let n1 = m[@coq_type_annotation] in
   n1
 
 let n2 =